fix: Create notification issues for manual release failures
On the one hand we may think that if someone is running this manually,
they'll notice it failing on their own; on the other hand, we create
issues for patch release failures (which are also triggered somewhat
manually), so at least this way we're consistent. Alternatively we could
remove the issue creation from the patch workflow and only create the
issue for periodic failures.